Se agregaron los modulos para que pueda almacenar fisicamente las bases de datos, asi como mejoramiento de los procedimientos de evaluacion de consultas

+++ src/examen1/Servidor.java
@@ -1,20 +1,82 @@
 package examen1;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.net.ServerSocket;
 import java.net.Socket;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.Set;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import java.util.*;
 
 public class Servidor {
+  public static void serializeDataOut(LinkedList table, String filepath) throws IOException{
+    FileOutputStream fos = new FileOutputStream(filepath);
+      try (ObjectOutputStream oos = new ObjectOutputStream(fos)) {
+          oos.writeObject(table);
+      }
+  }
+
+  public static LinkedList serializeDataIn(String filepath) throws IOException, ClassNotFoundException{
+   FileInputStream fin = new FileInputStream(filepath);
+   LinkedList table;
+      try (ObjectInputStream ois = new ObjectInputStream(fin)) {
+          table = (LinkedList) ois.readObject();
+      }
+   return table;
+ }
+      
+  public static HashMap<String, HashMap> loadFromData(){
+      String path = "data/";
+      HashMap<String, HashMap> bases = new HashMap<>();      
+      File loader = new File(path);
+      
+      for(File database: loader.listFiles()){
+         if(database.isDirectory()){
+            if(database.list().length > 0){   
+              HashMap<String, LinkedList> tables = new HashMap<>(); 
+              for(File table: database.listFiles()){
+                  try{
+                    tables.put(table.getName(), serializeDataIn(path + database.getName() + "/" + table.getName()));
+                  }
+                  catch(Exception e){
+                      e.printStackTrace();
+                  }
+              }
+              bases.put(database.getName(), tables);
+           }
+            else{
+                bases.put(database.getName(), null); //En caso de que la base de datos no tenga tabla alguna
+            }
+         }
+      }
+      return bases;
+  }
+
+  public static void saveDBToData(HashMap<String, HashMap> bases){
+      String path = "data/";
+      File loader = new File(path);     
+      Collection<String> oldbases = new ArrayList(Arrays.asList(loader.list()));
+      Collection<String> newbases = new ArrayList<>(bases.keySet());
+      
+      List<String> destinationList = new ArrayList<>(newbases);
+      destinationList.removeAll(oldbases);
+      
+      for(int i=0; i < destinationList.size(); i++){
+          File base = new File(path + destinationList.get(i));
+          try{
+              base.mkdir();
+          }
+          catch(Exception e){
+              e.printStackTrace();
+          }     
+      }
+  }
+    
+    
   public static void main(String[] args) {
-    HashMap <String, HashMap> bases = new HashMap<>();
+    HashMap <String, HashMap> bases;
     String baseActual = null;
     String dbName;
     
@@ -28,33 +90,48 @@ public class Servidor {
         System.out.println("Cliente conectado desde: " + cl.getInetAddress());
         for (;;) {
           ObjectInputStream ois = new ObjectInputStream(cl.getInputStream());
-          String query          = (String) ois.readObject();
-          query                 = query.toUpperCase();
+          String query                = (String) ois.readObject();
+          query                          = query.toUpperCase();
           
           /* Se checa que sentencia es la que se quiere */
           Sentencias se = new Sentencias();
           int a = se.getTipoSentencia(query);    
           String mensajeAEnviar = "";
           
+          /* Invocamos al metodo estático para cargar las bases de datos desde nuestro metadatos */
+          bases = loadFromData();
+          
           switch(a) {
             case 0:   /* Se crea la base de datos -> se pone en el hash map */
               if(se.verifySyntaxCreateDatabase(query)){
-                  dbName = se.createDatabase(query);                         // Se obtiene el nombre de la base con la que se hará la operación. 
-                  HashMap <String, LinkedList> base = new HashMap<>();                  // Se crea una instancia de HM, que contentra nombre de tabla y Lista de Objetos
-                  bases.put(dbName, base);                                          // Se añade al HashMap bases los datos obtenidos anteriormente.
-                  mensajeAEnviar = "Base de datos creada satisfactoriamente";       // Se envía un mensaje de que se creo de manera correcta.
+                  dbName = se.createDatabase(query);                                            // Se obtiene el nombre de la base con la que se hará la operación. 
+                  if(bases.containsKey(dbName)){
+                      mensajeAEnviar = "Error de Creación:  La base de datos ya existe";
+                  }
+                  else{
+                    HashMap <String, LinkedList> base = new HashMap<>();             // Se crea una instancia de HM, que contentra nombre de tabla y Lista de Objetos
+                    bases.put(dbName, base);                                                             // Se añade al HashMap bases los datos obtenidos anteriormente.
+                    saveDBToData(bases);
+                    mensajeAEnviar = "Base de datos creada satisfactoriamente";       // Se envía un mensaje de que se creo de manera correcta.
+                  }   
               }
               else{
-                  System.out.println("Error de Sintaxis: Create Database");
+                  mensajeAEnviar = "Error de Sintaxis: Create Database";
               }
               break; 
             
             case 1:  /* Se remueve la base de datos del hash map */
-              dbName = se.dropDatabase(query);                                  // Se obtiene el nombre de la base con la que se hará la operación.
+              if(se.verifySyntaxDropDatabase(query)){
+                dbName = se.dropDatabase(query);                                  // Se obtiene el nombre de la base con la que se hará la operación.
               if (bases.remove(dbName) == null) {                               // Si es nulo quiere decir que no hay esa llave; por ende no existe esa base.
                 mensajeAEnviar = "No existe una base de datos con ese nombre";
               } else {                                                          // Se eliminó bien !.
                 mensajeAEnviar = "Base eliminada satisfctoriamente!";
+              }  
+              
+              }
+              else{
+                  mensajeAEnviar = "Error de Sintaxis: Drop Database";
               }
               break;
